Patch 11 from Steves lazy parsing patchset.
Cache checksum for non-empty messages.

VersionMessage and AddressMessage require some special handling.  VersionMessage because it's never lazy parsed or cached.  AddressMessage because when serializing PeerAddresses the time field is dynamic.

Checksum byte array is currently transient so no gains for a message extracted from java serialization then bitcoinSerialized.  I don't think this would ever happen in real life but if it does then it could also be included in the serialized object.