Wallet: delete some code related to encrypted wallets.
Remove a c'tor that built an empty wallet that had a keycrypter but no password (so is it really encrypted?!), the sole purpose of which was to support another method that let you generate a new random key after the wallet is encrypted by providing the password. Post HD wallets doing this should be rare it's acceptable to require the user to manually encrypt the key before giving it to the wallet class.

This simplifies what's coming up.