Wallet: keychain migration, part 1
Switch the existing code to using only a single BasicKeyChain. Rework the encryption API's a bit as part of doing this. Disable key rotation for now as it will have to change in the HDW world anyway.

Functionality wise, this is largely a no-op, though BasicKeyChain should be faster at handling lots of keys at once.