Wallet: keychain migration, part 2
Expose the basic keychain object, make serialization/deserialization delegate to it. BasicKeyChain is now handling all of the tasks that Wallet[ProtobufSerializer] was doing itself.

Tweak the extensions API so that you no longer have to create a partly constructed Wallet before deserialisation: partly constructed wallets are not a good idea!

+ misc domino-effect changes that resulted from the above work.