Insert sigs in scriptSig in order of the signers
This way the first sig will be the one from LocalTransactionSigner, the
second sig will be the one from the next signer added to the wallet and so
on.
To achieve this, following keys should be added to the wallet in the same
order as signers. Pubkeys aren't sorted anymore when creating redeem
script, so now order of the keys matter. The first pubkey in redeem script
will always be the pubkey from the local (followed) keychain.