Update DefaultRiskAnalysis to consider IsStandard() Transaction Sizes
I noticed this class should give the user a warning when accepting
transactions - and it doesn’t consider the size of the transaction.
Only place I could see the size considered was inside
transaction.complete() which is when the user creates their own
transactions. Hope I committed this right, don’t hurt me.