Another max coins check, involving a series of transactions
Transactions that would cause estimated balance to exceed the maximum allowed are considered risky. There is
a risk that the culprit is actually a previous transaction and we're punishing the wrong one, but this
situation will clear up as transactions are confirmed. Most important is that the balance will not rise too
high.