Wallet, KeyChainGroup: Remove implicit creation and activation of a new chain.
The was likely the remains of lazy creation of wallets. Lazy creation wasn't of much use,
since wallets are usually accessed the moment after they're created. And also we introduced
a second "laziness" by only deriving a couple of keys per chain, deferring the hard work
for when bloom filtering is used.