kill ScriptChunk.startLocationInProgram
The startLocationInProgram field of ScriptChunk was used as a cacheing mechanism for storing the location of each ScriptChunk in a Script. This was memory-inefficient, as this integer was only actually used during interpretation of OP_CODESEPARATOR. Furthermore, it was dangerous and inconsistent: the ScriptBuilder methods, for example, do not properly set startLocationInProgram, leading to a potential runtime failure when interpreting ScriptBuilder-constructed Scripts.

Rather than cacheing each chunk's location, we add a method called size() that efficiently returns the size of a serialized ScriptChunk. During Script interpretation, we use this method to track the program counter.

Additionally, we add an exhaustive test to ensure that size() properly returns the size of a serialized ScriptChunk.

This will break any code that relied on ScriptChunk.getStartLocationInProgram().