MessageSerializer: Move protocolVersion from Message to here.
This change moves protocolVersion from Message to MessageSerializer.
The main reason is to support parsing transactions as non-segwit.
This is impossible to do otherwise, as MessageSerializer#makeTransaction()
does not accept a protocolVersion parameter.
Additionally this change makes the MessageSerializer similar to the
Stream in Bitcoin Core, where the protocol version can be modified.

 Conflicts:
	core/src/main/java/org/bitcoinj/core/BitcoinSerializer.java
	core/src/test/java/org/bitcoinj/core/BitcoinSerializerTest.java