Use protocolVersion in MessageSerializer
This change moves protocolVersion from Message to MessageSerializer.
The main reason is to support parsing transactions as non-segwit.
This is impossible to do otherwise, as MessageSerializer#makeTransaction
does not accept a protocolVersion parameter.
Additionally this change makes the MessageSerializer similar to the
Stream in BitcoinCore, where the protocol version can be modified.