PrefixedChecksummedBytes hierarchy: Add Comparable interface to Address, remove from PrefixedChecksummedBytes.
Requires address subclasses to implement compareTo() and provide the compareAddressPartial() method for comparing the first two fields.

This changes the natural ordering of addresses, and removes the natural ordering entirely for other PrefixedChecksummedBytes subclasses.

This also fixes a compareTo() collision regarding P2SH and non-P2SH addresses with the same bytes[].