Reorganize NetworkParameters subclass for easy diffs
Reorganize the initialization of the constants in each NetworkParameters
subclass so we can easily do diffs of `MainNetParms` vs `TestNet3Params`
and easily see how they are the same and how they are different.

I also reorganized the import statements so they “diff” better as well.

Other than those changes the only code I actually changed is
using the constant `PAYMENT_PROTOCOL_ID_UNIT_TESTS` instead of “unittest”
in getPaymentProcolId in UnitTestParams.

This will help us with the next step which is to create symbolic constants
for some of the parameters for the genesis blocks.