core: switch from BouncyCastle bcprov-jdk15to18 to bcprov-jdk15on
This will pave the way for switching to the bcprov-jdk18on JARs once
they make it to Maven Central. The build and tests seem to work,
but @schildbach can you try it on Android to make sure that your
Android build tooling doesn’t have any issues with multi-release jars?