Address/PrefixedChecksummedBytes: Use Network not NetworkParameters
This is one path towards moving Address to `o.b.base`. (Another path is to
reduce dependencies in `NetworkParameters` and move it to `o.b.base`)

This gets really messy because there are unit tests that assume a
UNITTEST NetworkParameters is stored inside Address (and other types)

I believe this needs to be fixed at some point, but it is not an easy change.