NetworkParameters: Use Network enum
This change makes NetworkParameters use and depend upon Network instead
of the reverse. This will allow Network to be moved to `base`

Define `UnitTestParams` as being on the `TEST` network. We must choose
a `Network` property for `UnitTestParams` and I have learned through trial-and-error
that `TEST` is the correct choice.

Not addressed in this PR is support for non-Bitcoin networks.