Address: Implement PrefixedChecksummedBytes internally
In preparation for future refactoring of Address, we don't want Address
tied to a base class with dependencies on o.b.core or NetworkParameters.
Instead we provide the NetworkParametersSupplier interface.

We also want the option of making Address itself an interface in the future.

I made PrefixedChecksummedBytes implement NetworkParametersSupplier, though
this was not strictly necessary. This will remind us to consider similar refactoring
for other subclasses of PrefixedChecksummedBytes. Yes, this results in some
code duplication but there is no need or use of polymorphism in the subclasses
of PrefixedChecksummedBytes.

Since equals() was copied from NetworkParameters into Address, create AddressTest
and copy the equals() test from PrefixedChecksummedBytesTest.