WIP/Proof-of-concept: Address as an interface
This needs additional work, especially the creation of some kind of AddressFactory interface,
but refactoring to an Address interface and getting unit tests to pass turned out
to be easier than expected.

Comments welcome, but please remember this is just a proof-of-concept and will be followed
by a more serious PR at a later time.