Address: implement PrefixedChecksummedBytes inline, rather than extend it
In preparation for future refactoring of Address, we don't want Address
tied to a base class with dependencies on o.b.core or NetworkParameters.

We also want the option of making Address itself an interface in the future.

This results in some code duplication but there is no strong need for or
use of polymorphism in the subclasses of PrefixedChecksummedBytes.

Since equals() was copied from NetworkParameters into Address, create AddressTest
and copy the equals() test from PrefixedChecksummedBytesTest.