ECKey: add BIP 137 support for signing text messages with segwit addresses
BIP 137 defined new header byte values for the created signature
when the signature was created with a private key from a segwit
address.

This commit extends the "signMessage(..)" with a new ScriptType argument
(because the value of the signature header since BIP 137 depends on the
address type). This commit keeps the old methods (without the ScriptType
argument), but marks them as deprecated (using P2SH as default, if called
without ScriptType, thus not changing old behaviour).