Wallet: use Instant and Instant.MIN to privately represent vKeyRotationTimestamp
This is a first step in an experiment to use Instant an Instant.MIN to represent
vKeyRotationTimeStamp internally in the wallet.

Where previously the value 0L was used to represent rotation not enabled,
we are now using Instant.MIN to represent the same thing. As a temporary solution,
the method vKeyRotationTimestampSeconds() can be used to get the backward-compatible
`long` before calling other methods that have not been converted yet.

By using MAX instead of 0 (EPOCH) we likely prevent that any key compared will ever
be rotated (even if the key has an unknown timestamp, e.g. 0/EPOCH)