TimeUtils: Use AtomicReference<Instance) internally
Doing some refactoring, I was getting false-positives (failures that should have succeeded)
in tests that were comparing actual vs. expected `Instant` values. The problem is that
as currently implemented `mockTime` only has millisecond resolution, but `Instant` has higher
resolution.  (I assume these failures were occurring with a `null` mockClock)

I believe that using `Instant` internally will fix these problems.

The `AtomicReference` was necessary for the "roll" functionality.

I think there is small race condition in rollMockClockMillis(long) between
the check for null (which throws IllegalStateException if null) and
the `updateAndGet`. I'm not sure if the correct exception will be thrown
if `updateAndGet` fails because of the `null`. That's why this is a DRAFT.