WIP: Step 0 of (de)serialization logic move from Message to BitcoinSerializer
Introduces a wrapper for payload bytes and a cursor position for reading,
and an offset.

* Replaces the payload, cursor and offset fields in Message
* All the "reading from payload" helper methods move from Message to here
* Reduces the amount of constructors in the Message hierarchy:
  previously there were often the permutation "payload" and "payload, offset"
* Should make it easier to move parse() from Message to BitcoinSerializer

The Payload class is missing many of our quality marks (JavaDocs and tests,
just to name a few). For now, this PR is only to test the concept.

Sadly, as with the other PR, it fails unit tests so it's draft for now.