Message: remove `length` and `optimalEncodingMessageSize` fields
This does away with a lot of updating the fields while parsing and
mutating the data.

`getMessageSize()` now returns the length of the data as
serialized by `bitcoinSerializeToStream()`, which can be different
from the length of an original payload.

The default implementation of `getMessageSize()` is not performant
enough for all usecases, to it is overridden in strategic subclasses
to calculate the size in a much more efficient way.