VarInt: fix long values accepted as an int value
`-1` and `Integer.MIN_VALUE` (and all values inbetween) are in fact
very large VarInts that can only be expressed as a long. So their
test data have been moved from `VarIntTest.integerTestVectors` to
`longTestVectors`.

`intValue()` is changed to guard against such cases, and a helper
`fitsInt()` is introduced.

JavaDocs are added to `intValue()` and `longValue()` to clarify the
expected value ranges.