POC/WIP: Split WalletTool into 2 classes
This is a proof-of-concept for splitting the WalletTool into 2 classes:

1. WalletTool: args parsing with CLI, help/documentation, stdout/stderr, exit (with code)

2. WalletToolService: Wraps everything about wallet/peerGroup/store state into
 a single object and methods. Error code/messages returned by throwing (this is
the best way to do it with pico-cli, I think)

This will not compile, but is an outline for how things can be structured. I carefully went through all the code in WalletTool and figured out what the methods should look like.

I just wanted to share this as it might be helpful to anyone working
on the WalletTool.