AddressParser: simplify by replacing Strict with Simple
These changes simplify the interface and change the names to more closely
match the use-cases we've been seeing internally and externally with alpha1.

parseAddressAnyNetwork() was both verbose and misleading.
AddressParser.Strict was being used (for example in the JavaFX wallet) more
as a context-specific simplification of the parent interface than as a "strict"
implementation.

parseAddressAnyNetwork() is deprecated as is the Strict sub-interface. They
can be removed before 0.17 final (after one more pre-release)