Wallet: wrap more exceptions as "failedFuture"
I'm pretty sure this is the right thing to do for `CompletionException`.
I'm less sure for `IllegalArgumentException` and `IllegalStateException`.

The `CompletionExceptions` cases need to be handled at runtime by wallet
implementations and should show up as failed futures (so that callers don't
have to test whether the future failed _and_ do a try-catch) for the "fatal"
exceptions I'm not as sure.