Remove deprecated AbstractTimeoutHandler
This was first deprecated in 0.17-alpha1 (i.e. it has is not deprecated
in the 0.16.x releases) so we might want to wait until
after 0.17 final is released before merging this PR.

However, this is fairly internal to the implementation so we might
be able to remove it in the next 0.17 alpha release.