WIP: ForwardingService Java 17 version
* examples, integration-test: Gradle use JDK 17+
* ForwardingService: Use Java 17 to simplify

Simplification:

* Move more code back to main method
* Use record class to collect config params
* static parse(args) method to build config
* Construct ForwardingService with Wallet and Config objects
* Use effectively final objects in the main try-with-resources
* Every variable -- except two locals in parse() -- is either final
  or effectively final.
* Keep important stuff towards the top of the file, details at
  the bottom.

The overall objective of this refactoring was to make the overall
flow clearly visible in the main method of the app while preserving
ForwardingService as a "reusable" and `Closable` object. I realize
that things could be further simplified by moving more code into
main, but I want this example to encourage creating modular "services"
with bitcoinj.