ForwardingService: rewrite forward() to be send()
Rewrite forward() to be a more generic send() method (which is essentially
just a wrapper to wallet.sendTransaction() that logs various completion states.

Note that we moved the whenComplete blog that logs "relayed" status so users
can use this code in more cases. This is because a common question that
comes up is "why was my transaction not seen?". So making this logging
code more reusable should be helpful.