graalvm.yml: Use setup_java GitHub Action
We can now use the same setup_java GitHub Action
in gradle.yml and graalvm.yml. This commit switches
to the setup_java action from the setup-graalvm action
and also changes some variable names to more closely
match gradle.yml.

(It might even be possible/desirable to combine the
two workflows in the future, but let's keep this PR simple)
# Conflicts:
#	.github/workflows/graalvm.yml