wallettemplate: add Launcher class
This works-around a limitation in JavaFX that can prevent non-modular
JavaFX classes from launching properly.

With this change the following should work:

gradle bitcoinj-wallettemplate:installDist
./wallettemplate/build/install/bitcoinj-wallettemplate/bin/bitcoinj-wallettemplate

It will likely also allow running the app in IntelliJ without using Gradle 

This change is unnecessary for a modular JavaFX and can be reversed
when/if we upgrade wallettempalte to a modular app.