base: remove last @Nonnull annotations & findbugs dependency
A previous PR removed the last @Nullable annotation.

This PR removes the @Nonnull annotations as we can now assume that
everything in this module is non-nullable. If we discover otherwise
we can add JSpecify and mark them with that. We may want to use
JSpecify anyway, but I'll leave that for a separate PR.