modular-architecture.md: show slf4j & JSpecify as deps of base
This is just a cosmetic/informational change. In the interest of keeping
the diagram simple, I have not been showing multiple links to minor direct
dependencies. But this change chooses to display slf4j & JSpecify as
direct dependencies of `base`. This makes it more clear that
`base`, `core` and `wallet` all will depend on them. The diagram
still "lies" a little bit in not showing that they are also direct
dependencies, but I think that is an acceptable tradeoff for a diagram
that is easier to read.