modular-architecture.md: show SLF4J and JSpecify as dependencies of base
This is just a informational change. In the interest of keeping
the diagram simple, I have not been showing multiple links to minor direct
dependencies. But this change chooses to display SLF4J & JSpecify as
direct dependencies of `base`. This makes it more clear that
`base`, `core` and `wallet` all will depend on them. The diagram
still "lies" a little bit in not showing that they are also direct
dependencies, but I think that is an acceptable tradeoff for a diagram
that is easier to read.