TestNet3Params: fix difficulty transition check
The difficulty transition check didn't properly account for the
"20 minute rule" specific to testnet: if the time delta between
a block and its previous block is more than 20 minutes, the block's
difficulty target MUST reset to the minimum difficulty.

Before this fix, it did not check the difficulty target at all in
this case. So it wrongly assumed the difficulty target can be changed
to any value.

This also extracts a private checkDifficultyTarget() method to make
the logic easier to read.