Block: remove exception-safe usage mode from `checkProofOfWork()`
This removes both the boolean argument and the boolean return type.
It always throws if difficulty target is not met. If it returns
normally, the target is met.

The only consumer of the exception-safe usage mode is switched to
using `Difficulty.isMetBy()` directly.

Also, this method is switched to private visibility.