FakeTxBuilder: do not solve fake blocks
Because `solve()` in almost all cases requires too much work, in
future it should be called explicitly by the tests that depend on
proper proof of work.

Most tests already use `Context.relaxProofOfWork` to skip the
check. Only PeerTest depends on it, so it is changed to solve one
block manually (on easiest difficulty).

Some method names were stripped of the adjective "solved" and the
JavaDocs amended appropriately.