Experiment: Create SignedTransactionSubclass
This experiment shows that we can mostly assume that any transaction that
is deserialized via `read()` can be assumed signed and mostly immutable.

The "WalletTransaction" methods (to set purpose, exchange rate, and memo)
were left mutable in this experiment. A separate effort to create
a wallet transaction that stores mutable wallet annotated info can
address those fields.

Besides the changes to Transaction and BitcoinSerializer, the only other
changes needed were to test code (which mutates deserialized and roundtrip
transactions, so the `.asUnsigned()` conversion method needed to be called.