DeterministicKeyChain: Fix depth issue in toEncrypted()
In the DeterministicKeyChain constructor that uses a KeyCrypter to
encrypt an existing keychain, there was a bug where encrypted non-leaf keys
that are grandchildren or lower of the root were being constructed with
their parent set to root rather than their direct encrypted parent.

This commit fixes the issue.

The problem was undetected until PR #3709 added `depth` to DeterministicKey's
`equals` and `hashCode`, which was than detecting the incorrect depth
after a roundtrip.