HDPath: add TODO comment about asPartial only on HDFUllPath
Having asPartial() available on all HDPath subtypes discourages the use of
consistent strong typing. We should consider removing it at some point.

We can either merge this now to put the reminder in the code or leave
it open until we finish this round of refactoring.

I realize this could have been an Issue, but since we're focused on
PRs right now, I thought I'd put this in the list.