DeterministicKey: narrow field `childNumberPath` to type `HDPartialPath`
Use an `HDPartialPath` for the private field `childNumberPath`.
Convert to an `HDFullPath` in the getter based upon on the boolean
value returned by `isWatching()`.

We might decide to use an `HDPartialPath` in the canonical constructor
(and possibly others) if we want to encourage stricter usage of types,
but for this change it's out of scope.