DeterministicKey: Use HDPartialPath for internal childNumberPath
Use an HDPartialPath for the internal `private` field `childNumberPath`.
Convert to an `HDFullPath` in the getter based upon on the boolean
value returned by `isWatching()`.

We might decide to use an `HDPartialPath` in the canonical constructor 
(and possibly others) if we want to encourage stricter usage of types,
but that change would be in a separate PR.