Wallet: deprecate UTXOProvider support
Deprecate getUTXOProvider/setUTXOProvider and protected methods
that use the UTXOProvider methods.

We have plans to refactor and simplify the wallet interface and removing
this mode of operation (which is rarely used outside of tests) will
help us move in that direction.

Longer term, we hope to provide a more flexible wallet interface that
will make it even easier to use other sources of UTXO information.