integration-test, wallettemplate, wallettool: Use JUnit BOM in build.gradle
By including the BOM, Gradle will automatically make sure all JUnit
libraries use the same version. So we only need to specify the version
on the BOM `and the others will follow.

Use  of the BOM is required to get JUnit 5.12 to work with (at least
some versions of) Gradle.