TransactionInputTest: use getter in disconnect tests
In tests `testStandardWalletDisconnect()` and `testUTXOWalletDisconnect()`
use getter `getConnectedOutput()` instead of the `connectedOutput` member.

Note that this assert

 assertNull(txOutPoint.connectedOutput);

changes to

 assertEquals(tx1.getOutput(txOutPoint.index()), txOutPoint.getConnectedOutput());

This reflects the fact that `getConnectedOutput()` will find the connected
output via `fromTx`. This changes the test slightly in order to reflect
the public behavior of `getConnectedOutput()` rather than the internal
state of the `connectedOutput` member.