PeerGroup: replace use of Guava Ordering with a Stream
Instead of reverse sorting with Guava Ordering and then counting with
a for loop, use a Stream and Collectors.counting() to do the counting.
Then convert from Map.Entry to Pair and sort by `count`.

The Pair class is kept, but simplified to a record-like class. Rather
than have a comparator that wouldn't match equals, we'll use a lambda
comparator in the `.sorted()` stream operation.