TransactionOutPoint: {connect,disconnect}Transaction() set connectedOutput to `null`
In connectTransaction() set `connectedOutput` to `null` in the new copy that is returned.
We do this because `connectedOutput` is unused when `fromTx` is non-null.

In disconnectTransaction() set `connectedOutput` to `null` by calling disconnectOutput().
We do this because a disconnected TransactionOutPoint should be fully disconnected.

A subsequent commit requires that `connectedOutput` and `fromTx` aren't both non-null.