core: Upgrade to Guava 33.4.8
Guava 33.4.8 is a modular Guava. To get wallettemplate to build,
we needed to add a `requires com.google.common` to its `module-info.java`
and we also had to remove a dependency constraint in core/build.gradle.

Removing that dependency constraint may cause issues on Android, so
this commit/branch/PR should be considered WIP, for now.