WalletProtobufSerializer: simplify writeConfidence() parameters
Instead of passing TransactionConfidence & Protos.TransactionConfidence.Builder
objects to writeConfidence(), just pass the Transaction and create them
inside the method. They are temporary objects and not used anywhere else.

This makes the code a little more readable and it also eliminates an
IDE warning about synchronizing on `confidence`.

Also add some explanatory comments and a suggestion for further refactoring.