WIP: Support for building with Gradle 9
This is work-in-progress and currently comments out the
wallettemplate and integration-test modules to avoid issues
in building those modules with Gradle 9.

Also, some of the other changes I made may not be the best way of resolving
issues.

I didn't spend much time trying to resolve those issues, so fixing
them may not be hard. This was just a quick test to get a feel for
how difficult Gradle 9 support will be for us.

GHA gradle.yml: Use Gradle 8.3 rather than 8.0 in matrix

Re-enable wallettemplate with BA jlink 3.1.2 and min Gradle 8.11.1

Re-enable integration-test

GHA gradle.yml: add back wallettemplate installDist and jlink targets

# Conflicts:
#	.github/workflows/gradle.yml
#	wallettemplate/build.gradle