build.gradle: improve Guava Android constraint
Instead of forcing usage of the Android version of Guava on the
core `implementation` configuration, do it on the `testImplementation`
configuration of `core` and the `implementation` configuration of
`integration-test`.

This makes sure our tests run against the Android version, which is
what we really want. But with this change we can use the JRE version
in other modules, such as the `wallettemplate` module.