build.gradle: parameterize minimum JDK to test on
Rather than use a boolean that runs hardcoded JDK8 tests, use
a numeric parameter to set the minimum JDK to test on. Use
`-PtestMinJdk=8` instead of `-testJdk8=true`.

This makes it easier to change our minimum version and it also makes
it possible to run the tests on the minimum available JDK on a particular
platform or really any JDK you want.