gradle.yml, .gitlab-ci.yml: run `installDist` target on every subproject
`installDist` is a no-op on subprojects that don't use the `application`
plugin, but runs on every subproject that does: currently wallettool and
wallettemplate.

It also makes our CI configuration files simpler because it lets
Gradle decide what to run and we no longer have to make changes to CI
configuration files adding or removing `project:installDist` tasks on
Gradle invocations.

This change will increase test coverage in some configurations.